requirements.txt
pystray==0.19.5
Pillow==9.0.1
